Adds API call for now playing movies
Implements functionality to retrieve and display now playing movies from the Movie Database API.

The script now takes an argument to specify the type of movie information to retrieve. When the "playing" type is specified, the script calls the API, parses the JSON response, and prints the title of each movie.